+++ app/src/main/java/in/paperwrk/safetycollabproject/activities/SirenSoundActivity.java
@@ -1,18 +1,20 @@
-package in.paperwrk.safetycollabproject;
+package in.paperwrk.safetycollabproject.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 
-public class TrustedContactsActivity extends AppCompatActivity {
+import in.paperwrk.safetycollabproject.R;
+
+public class SirenSoundActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_trusted_contacts);
+        setContentView(R.layout.activity_siren_sound);
 
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-        setTitle("Trusted Contacts");
+        setTitle("Siren - Sound");
     }
 }
