+++ app/src/main/java/in/paperwrk/safetycollabproject/fragments/SOSFragment.java
@@ -1,4 +1,4 @@
-package in.paperwrk.safetycollabproject;
+package in.paperwrk.safetycollabproject.fragments;
 
 
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import in.paperwrk.safetycollabproject.R;
 
 
 /**
