+++ app/src/main/res/layout/activity_about.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AboutActivity">
+    tools:context=".activities.AboutActivity">
 
     <include
         android:id="@+id/toolbar"
