+++ app/src/main/res/layout/activity_signin.xml
@@ -24,7 +24,8 @@
             android:textColor="#757575"
             android:textSize="18sp" />
 
-        <android.support.v7.widget.CardView xmlns:card_view="http://schemas.android.com/apk/res-auto"
+        <android.support.v7.widget.CardView
+            xmlns:card_view="http://schemas.android.com/apk/res-auto"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="16dp"
