+++ app/src/main/res/layout/activity_siren_sound.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SirenSoundActivity"
+    tools:context=".activities.SirenSoundActivity"
     android:orientation="vertical">
 
     <include android:id="@+id/toolbar" layout="@layout/layout_toolbar" />
