+++ app/src/main/res/layout/activity_trusted_contacts.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".TrustedContactsActivity">
+    tools:context=".activities.TrustedContactsActivity">
 
     <include
         android:id="@+id/toolbar"
