+++ app/src/main/res/layout/fragment_fake_call.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_light"
-    tools:context=".FakeCallFragment">
+    tools:context=".fragments.FakeCallFragment">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
