+++ app/src/main/res/layout/fragment_sos.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_light"
-    tools:context=".SOSFragment">
+    tools:context=".fragments.SOSFragment">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
