+++ app/src/main/res/layout/fragment_track_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TrackUserFragment">
+    tools:context=".fragments.TrackUserFragment">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
