+++ app/src/main/res/layout/page_intro_activity4.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PageIntroActivity"
+    tools:context=".activities.PageIntroActivity"
     android:background="@color/screen4">
     <LinearLayout
         android:layout_width="wrap_content"
